Initial Project Commit
This projected was first created in 2015. This is a Brick Breaker game
made in Unity with C#. It mostly served as a playground for testing and
learning Unity. There are a lot of improvements that can be made such as
adding further levels and optimization (removing objects that are no
longer in play, i.e. particle effects). The purpose of this commit is to
showcase the project at its current working state.